pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-check-repo"
version = "0.1.0"
description = "AI-powered repository evolution analyzer"
license = {text = "MIT"}
requires-python = ">=3.11"
readme = "README.md"
keywords = ["git", "ai", "claude", "code-review", "repository", "analysis"]

dependencies = [
    "claude-code-sdk>=0.0.3",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
claude-check-repo = "cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"